Added SignatureAlgorithmHeader enum and verification
Added a new enum `SignatureAlgorithmHeader` to the `SignatureVerificationMiddleware` class in the `Core.API.ResponseHandling` namespace. This enum currently includes `ED25519` as a value. A check has been implemented in the same class to verify if the `algorithmHeader` can be parsed into the `SignatureAlgorithmHeader` enum. If parsing fails, an error is logged and a custom error response with a `Forbidden` status code is returned. Additionally, a new header `x-algorithm` has been added to the config headers in the `requestInterceptor` function in `axios.ts`, indicating the algorithm used for the signature.